Adds sub verticalze()
Add a file with sub verticalize().  The subroutine inputs a Horizontal
table with N dimension columns, M measure/fact columns, and R rows, then
outputs a Vertical table with N+1 columns and M*R rows.  Purpose is to
create a table that can be easily summarized with a Pivot Table,
starting from a rectangular dataset.